Main

Documentation

type Coordenadas = (Int, Int)

type Mapa = [String]

type NBot = Int

type TimeLeft = Int

type Raio = Int

type DimensaoMapa = Int

main :: IO ()

Module : Main

Description : Módulo Haskell que contém a função que reage à passagem do tempo (função avanca).

Este módulo contém funções recursivas e não recursivas com o intuito de devolver um mapa atualizado consoante o efeito da passagem de um instante de tempo.

avanca :: Mapa -> TimeLeft -> Mapa

A função avanca devolve um novo estado de jogo consoante a passagem de um instante de tempo.

Exemplo de utilização:

>>> avanca m 44 = ["#########","####### #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?# # #","#  ??   #","#########","* 4 5 0 3 1","* 5 6 1 3 1","0 5 4","1 6 5"]

tempo :: Mapa -> TimeLeft -> DimensaoMapa -> Mapa

A função tempo devolve um novo estado de jogo consoante o tempo que falta para a espiral começar:

Se os instantes que faltam para o jogo acabar for superior aos instantes que faltam para a espiral começar, então devolve o próprio mapa;

Se os instantes que faltam para o jogo acabar for igual a 0 (ou seja, o jogo terminou), então devolve o próprio mapa;

Se os instantes que faltam para o jogo acabar for igual ou inferior aos instantes que faltam para a espiral começar, então, uma pedra cai na posição conveniente (com recurso à função novomapaComEspiral)

dimensaomapa :: Mapa -> DimensaoMapa

A função dimensaomapa dá a dimensão do mapa (consoante qualquer estado de jogo apresentado no input).

Exemplo de utilização:

>>> dimensaomapa m = 9

novomapaComEspiral :: Mapa -> TimeLeft -> Mapa

A funcao novomapaComEspiral devolve um novo estado de jogo com a continuação (ou início) do efeito espiral.

As funções veOndeCaiPedra e novomapaComEspiral1 são as principais funções auxiliares desta função.

veOndeCaiPedra :: Mapa -> Coordenadas -> Mapa

A função veOndeCaiPedra averigua se a coordenada onde a pedra (da espiral) caiu em cima de uma bomba, power up ou jogador, e devolve o mapa consoante o pretendido:

Se caiu em cima de uma bomba, essa bomba é removida do mapa (com recurso à função novomapaSemBombaExpl);

Se caiu em cima de um power up, esse power up é removido do mapa (com recurso à função novomapaSemPu);

Se caiu em cima de um (ou vários, se tiverem mais do que um na mesma célula) jogador, esse(s) jogador(es) perde(m) o jogo (com recurso à função novomapaSemJog);

Caso contrário, isto é, se simplesmente caiu em cima de um espaço vazio (sem power up), pedra, ou tijolo, devolve o próprio mapa.

novomapaComEspiral1 :: Mapa -> TimeLeft -> Mapa

A função novomapaComEspiral1 devolve o mapa com a pedra (da espiral) colocada no sítio correto.

Esta função recorre a algumas funções auxiliares, entre as quais novomapaComEspiral2 e novomapaComEspiralAux.

novomapaComEspiral2 :: Mapa -> DimensaoMapa -> DimensaoMapa -> Int -> Int -> TimeLeft -> Mapa

novomapaComEspiralAux :: Mapa -> Mapa -> Coordenadas -> Coordenadas -> Mapa

novaLinhaP :: Mapa -> Coordenadas -> String

A função novaLinhaP, com o auxílio das funções substituiPedra e pegaLinha, dado a mapa do estado atual e as coordenadas onde cai a pedra, dá como resultado a nova linha com a pedra (#) na posição desejada.

substituiPedra :: String -> Coordenadas -> String

Na função substituiPedra, dada a linha do mapa onde cai a pedra e a respetiva coordenada, devolve a linha com a pedra (#) na posição desejada.

coordenadas :: TimeLeft -> DimensaoMapa -> DimensaoMapa -> Int -> Int -> Coordenadas

A função coordenadas, dá as coordenadas em que a próxima pedra vai cair.

Esta função tem como parâmetros:

t: número instantes de tempo que faltam para o jogo terminar;

d , s: dimensão do mapa (estes valores encontram-se duplicados porque para as funções auxiliares é necessário que eles se encontrem desta forma);

i1: número 1 (este 1 é necessário para as funções auxiliares, é uma espécie de "contador");

i2: número 2 (este 2 é necessário para as funções auxiliares, é uma espécie de "contador").

Exemplos de utilização:

>>> coordenadas 44 9 9 1 2 = (6,1)

>>> coordenadas 5 13 13 1 2 = (7,7)

coordenadasAux :: TimeLeft -> [Coordenadas] -> Coordenadas

A função coordenadasAux auxilia a função coordenadas.

contadorEspiral :: DimensaoMapa -> DimensaoMapa -> Int -> Int -> [Coordenadas]

A função contadorEspiral dá como resposta as coordenadas do mapa pela ordem em que a espiral cai.

Exemplos de utilização:

>>> contadorEspiral 9 9 1 2 = [(1,1),(2,1),(3,1),(4,1),(5,1),(6,1),(7,1),(7,2),(7,3),(7,4),(7,5),(7,6),(7,7),(6,7),(5,7),(4,7),(3,7),(2,7),(1,7),(1,6),(1,5),(1,4),(1,3),(1,2),(2,2),(3,2),(4,2),(5,2),(6,2),(6,3),(6,4),(6,5),(6,6),(5,6),(4,6),(3,6),(2,6),(2,5),(2,4),(2,3),(3,3),(4,3),(5,3),(5,4),(5,5),(4,5),(3,5),(3,4),(4,4)]

>>> contadorEspiral 5 5 1 2 = [(1,1),(2,1),(3,1),(3,2),(3,3),(2,3),(1,3),(1,2),(2,2)]

contadorX :: DimensaoMapa -> DimensaoMapa -> Int -> [Int]

A função contadorX, dá a lista das coordenadas X (ou seja, as abcissas) do mapa espiral pela ordem em que as pedras caiem.

Esta função tem o auxílio das seguintes funções: contadorX1, contadorX1A, contadorX2,contadorX2A, contadorX2AA, contadorX2AAA, contadorX3, contadorX4, contadorX4A, contadorX5, contadorX5A e contadorX6.

contadorX6 :: DimensaoMapa -> DimensaoMapa -> Int -> [[Int]]

contadorX5 :: DimensaoMapa -> Int -> [[Int]]

contadorX5A :: DimensaoMapa -> Int -> [[Int]]

contadorX4 :: DimensaoMapa -> DimensaoMapa -> [[Int]]

contadorX4A :: DimensaoMapa -> DimensaoMapa -> [[Int]]

contadorX3 :: DimensaoMapa -> Int -> [[Int]]

contadorX2 :: DimensaoMapa -> Int -> [[Int]]

contadorX2AAA :: [[Int]] -> [[Int]]

contadorX2AA :: DimensaoMapa -> Int -> [[Int]]

contadorX2A :: DimensaoMapa -> Int -> [[Int]]

contadorX1 :: DimensaoMapa -> Int -> [[Int]]

contadorX1A :: DimensaoMapa -> Int -> [[Int]]

contadorY :: DimensaoMapa -> DimensaoMapa -> Int -> Int -> [Int]

A função contadorY, dá a lista das coordenadas Y (ou seja, as ordenadas) do mapa espiral pela ordem em que as pedras caiem.

Esta função tem o auxílio das seguintes funções: contadorY1, contadorY1A, contadorY2,contadorY2A, contadorY3, contadorY4, contadorY4A, contadorY5, contadorY5A, contadorY5AA, contadorY5AAA, contadorX6 ordenacao e ordenacao2.

ordenacao2 :: [[Int]] -> [[Int]] -> [Int]

contadorY6 :: DimensaoMapa -> Int -> [[Int]]

contadorY5 :: DimensaoMapa -> Int -> [[Int]]

contadorY5AAA :: [[Int]] -> [[Int]]

contadorY5AA :: DimensaoMapa -> Int -> [[Int]]

contadorY5A :: DimensaoMapa -> Int -> [[Int]]

contadorY4 :: DimensaoMapa -> Int -> [[Int]]

contadorY4A :: DimensaoMapa -> Int -> [[Int]]

contadorY3 :: DimensaoMapa -> DimensaoMapa -> Int -> [[Int]]

ordenacao :: [[Int]] -> [[Int]] -> [[Int]]

contadorY2 :: DimensaoMapa -> DimensaoMapa -> [[Int]]

contadorY2A :: DimensaoMapa -> DimensaoMapa -> [[Int]]

contadorY1 :: DimensaoMapa -> Int -> [[Int]]

contadorY1A :: DimensaoMapa -> Int -> [[Int]]

eliminavazio :: [[Int]] -> [[Int]]

A função eliminavazio recebe uma lista de listas de inteiros, e retira dessa lista, as listas vazias.

Início Passagem de Tempo

chamasBombas :: Mapa -> [Coordenadas]

A função chamasBombas dá como resultado a lista das coordenadas que as chamas das bombas atingem.

Tendo em consideração que as chamas não atravessam:

As pedras;

Os tijolos;

Os espaços vazios com power up.

chamasBombasAux :: Mapa -> [Coordenadas] -> [[Raio]] -> [Coordenadas]

A função chamasBombasAux é a principal função auxiliar da função chamasBombas.

Esta função concatena o resultado das 4 primeiras (por vezes únicas, depende do raio de destruição da bomba) direções (cima, baixo, esquerda e direita) que as chamas percorrem; e tem como passo recursivo, percorrer as seguintes 4 direções.

tiraCoordPedras :: Mapa -> [Coordenadas] -> [Coordenadas]

A função tiraCoordPedras, tem como parâmetros o mapa e a lista das coordenadas que as chamas das bombas atingem, e retira dessa lista as coordenadas que correspondem a pedras.

eliminaChamasD :: Mapa -> Coordenadas -> [Raio] -> [Coordenadas]

A função eliminaChamasD, tem como parâmetros:

O mapa;

A coordenada de uma bomba que explode;

A lista do raio que essa bomba alcança (de 1 até ao raio da bomba).

Esta função pára de contabilizar as coordenadas das chamas das bombas a partir do momento em que encontra uma pedra, um tijolo, ou um espaço vazio com power up.

Dá como resultado as coordenadas que as chamas das bombas atingem, neste caso inferiores (para baixo) do local da bomba.

eliminaChamasU :: Mapa -> Coordenadas -> [Raio] -> [Coordenadas]

A função eliminaChamasU, tem como parâmetros:

O mapa;

A coordenada de uma bomba que explode;

A lista do raio que essa bomba alcança (de 1 até ao raio da bomba).

Esta função pára de contabilizar as coordenadas das chamas das bombas a partir do momento em que encontra uma pedra, um tijolo, ou um espaço vazio com power up.

Dá como resultado as coordenadas que as chamas das bombas atingem, neste caso superiores (para cima) do local da bomba.

eliminaChamasL :: Mapa -> Coordenadas -> [Raio] -> [Coordenadas]

A função eliminaChamasL, tem como parâmetros:

O mapa;

A coordenada de uma bomba que explode;

A lista do raio que essa bomba alcança (de 1 até ao raio da bomba).

Esta função pára de contabilizar as coordenadas das chamas das bombas a partir do momento em que encontra uma pedra, um tijolo, ou um espaço vazio com power up.

Dá como resultado as coordenadas que as chamas das bombas atingem, neste caso para a esquerda do local da bomba.

eliminaChamasR :: Mapa -> Coordenadas -> [Raio] -> [Coordenadas]

A função eliminaChamasR, tem como parâmetros:

O mapa;

A coordenada de uma bomba que explode;

A lista do raio que essa bomba alcança (de 1 até ao raio da bomba).

Esta função pára de contabilizar as coordenadas das chamas das bombas a partir do momento em que encontra uma pedra, um tijolo, ou um espaço vazio com power up.

Dá como resultado as coordenadas que as chamas das bombas atingem, neste caso __para a direita do local da bomba.

eliminaRepetidos :: [Coordenadas] -> [Coordenadas]

Na função eliminaRepetidos, dada uma lista de coordenadas, elimina dessa lista as coordenadas que se encontram repetidas. (Isso acontece quando bombas diferentes explodem e atingem as mesmas células).

eliminaCoord :: Coordenadas -> [Coordenadas] -> [Coordenadas]

A função eliminaCoord é a função auxiliar da função eliminaRepetidos.

chamasBombasTOTAIS :: Mapa -> [Coordenadas]

A função chamasBombasTOTAIS dá como resultado a lista das coordenadas que as chamas das bombas atingem, sem eliminar as coordenadas repetidas (deste modo, com esta função, conseguimos verficar se bombas diferentes atingem as mesmas células simultaneamente).

coincide :: Mapa -> Bool

A função coincide averigua se as chamas das bombas que explodem ao mesmo tempo atingem coordenadas em comum.

variasBombasExplodem :: Mapa -> Mapa

A função variasBombasExplodem dá como resposta o mapa, quando várias bombas explodem e atingem as mesmas células.

doIt :: Mapa -> [Coordenadas] -> Mapa

A função doIt tem como parâmetros:

O mapa;

A lista das coordenadas que as bombas que explodem atingem no próximo estado de jogo.

Com o especial recurso às seguintes funções: novomapaSemPI, novomapaSemJog, novomapaSemPU, novomapaSemBombaExpl e novomapaComInst1, esta função "faz explodir" o que está nas coordenadas que são atingidas pelas coordenadas das bombas.

eliminaBombasExpl :: Mapa -> [Coordenadas] -> Mapa

A função eliminaBombasExpl, elimina do mapa as informações das bombas que acabaram de explodir.

instante1 :: Mapa -> Coordenadas -> Bool

A função instante1 averigua, a partir de uma coordenada, se essa bomba vai explodir ou não no próximo instante de tempo.

borala :: Mapa -> Mapa

A função borala dado o mapa do estado atual do jogo, devolve o novo mapa atualizado tendo em conta se:

Alguma bomba explode no próximo instante de tempo;

Nenhum bomba explode no próximo instante de tempo.

analisa :: Mapa -> Mapa

A função analisa dado o mapa do estado atual do jogo, devolve o novo mapa atualizado, sabendo que há bombas a explodirem no próximo instante de tempo, e tem em conta se:

Há bombas a explodir em que as mesmas células são atingidas por diferentes bombas;

Todas as bombas que explodem não atingem células em comum.

prossegue :: Mapa -> Mapa

A função prossegue percorre o mapa, e faz explodir as bombas. Devolve o novo estado de jogo atualizado.

prossegue1 :: Mapa -> [[Raio]] -> [Coordenadas] -> Mapa

A função prossegue1 é a principal função auxiliar da função prossegue.

verificaCaminho :: Mapa -> [Raio] -> [Raio] -> [Raio] -> [Raio] -> Coordenadas -> Mapa

A função verificaCaminho averigua se existe algum jogador na coordenada da bomba.

Caso afirmativo, elimina-o do jogo;

Caso contrário, continua a analisar o caminho das chamas.

oqueaconteceD :: Mapa -> [Raio] -> [Raio] -> [Raio] -> [Raio] -> Coordenadas -> Mapa

A função oqueaconteceD tem como parâmetros:

O estado atual do jogo;

Lista do raio da bomba (4x) (para poderem ser explorados por funções diferentes, intactos);

Coordenadas da bomba em causa (que está a explodir).

Esta função vai auxiliar no processo de explosão de uma bomba, no que diz respeito às coordenadas para baixo da mesma.

oqueaconteceU :: Mapa -> [Raio] -> [Raio] -> [Raio] -> [Raio] -> Coordenadas -> Mapa

A função oqueaconteceU tem como parâmetros:

O estado atual do jogo;

Lista do raio da bomba (4x) (para poderem ser explodrados por funções diferentes, intactos);

Coordenadas da bomba em causa (que está a explodir).

Esta função vai auxiliar no processo de explosão de uma bomba, no que diz respeito às coordenadas para cima da mesma.

oqueaconteceL :: Mapa -> [Raio] -> [Raio] -> [Raio] -> [Raio] -> Coordenadas -> Mapa

A função oqueaconteceL tem como parâmetros:

O estado atual do jogo;

Lista do raio da bomba (4x) (para poderem ser explodrados por funções diferentes, intactos);

Coordenadas da bomba em causa (que está a explodir).

Esta função vai auxiliar no processo de explosão de uma bomba, no que diz respeito às coordenadas para a esquerda da mesma.

oqueaconteceR :: Mapa -> [Raio] -> [Raio] -> [Raio] -> [Raio] -> Coordenadas -> Mapa

A função oqueaconteceR tem como parâmetros:

O estado atual do jogo;

Lista do raio da bomba (4x) (para poderem ser explodrados por funções diferentes, intactos);

Coordenadas da bomba em causa (que está a explodir).

Esta função vai auxiliar no processo de explosão de uma bomba, no que diz respeito às coordenadas para a direita da mesma.

propriopormenor :: Coordenadas -> Mapa -> String

A função propriopormenor vê com mais pormenor, o que está na própria coordenada.

Nesta função apenas avaliamos se essa coordenada tem algum jogador ou não.

espacoVazioPP :: Coordenadas -> Mapa -> String

A função espacoVazioPP é função auxiliar da função propriopormenor.

downpormenor :: Coordenadas -> [Raio] -> Mapa -> String

A função downpormenor vê com mais pormenor, o que está x posições abaixo de uma determinada coordenada.

Respostas possíveis:

Pedra;

Tijolo;

Power up;

Bomb;

Jogador;

Nada.

espacoVazioD :: Coordenadas -> [Raio] -> Mapa -> String

A função espacoVazioD é a principal função auxiliar da função downpormenor.

uppormenor :: Coordenadas -> [Raio] -> Mapa -> String

A função uppormenor vê com mais pormenor, o que está x posições acima de uma determinada coordenada.

Respostas possíveis:

Pedra;

Tijolo;

Power up;

Bomb;

Jogador;

Nada.

espacoVazioU :: Coordenadas -> [Raio] -> Mapa -> String

A função espacoVazioU é a principal função auxiliar da função uppormenor.

leftpormenor :: Coordenadas -> [Raio] -> Mapa -> String

A função leftpormenor vê com mais pormenor, o que está x posições à esquerda de uma determinada coordenada.

Respostas possíveis:

Pedra;

Tijolo;

Power up;

Bomb;

Jogador;

Nada.

espacoVazioL :: Coordenadas -> [Raio] -> Mapa -> String

A função espacoVazioL é a principal função auxiliar da função leftpormenor.

rightpormenor :: Coordenadas -> [Raio] -> Mapa -> String

A função rightpormenor vê com mais pormenor, o que está x posições à direita de uma determinada coordenada.

Respostas possíveis:

Pedra;

Tijolo;

Power up;

Bomb;

Jogador;

Nada.

espacoVazioR :: Coordenadas -> [Raio] -> Mapa -> String

A função espacoVazioR é a principal função auxiliar da função rightpormenor.

downc :: Coordenadas -> [Raio] -> Coordenadas

A função downc diz qual é a primeira coordenada do "fogo" da bomba (para baixo do bomba).

upc :: Coordenadas -> [Raio] -> Coordenadas

A função upc diz qual é a primeira coordenada do "fogo" da bomba (para cima do bomba).

leftc :: Coordenadas -> [Raio] -> Coordenadas

A função leftc diz qual é a primeira coordenada do "fogo" da bomba (para a esquerda do bomba).

rightc :: Coordenadas -> [Raio] -> Coordenadas

A função rightc diz qual é a primeira coordenada do "fogo" da bomba (para a direita do bomba).

down :: Coordenadas -> [Raio] -> Mapa -> Char

A função down diz o que está abaixo da bomba que vai explodir (#,' ' ou ?).

up :: Coordenadas -> [Raio] -> Mapa -> Char

A função up diz o que está acima da bomba que vai explodir (#,' ' ou ?).

left :: Coordenadas -> [Raio] -> Mapa -> Char

A função left diz o que está à esquerda da bomba que vai explodir (#,' ' ou ?).

right :: Coordenadas -> [Raio] -> Mapa -> Char

A função right diz o que está à direita da bomba que vai explodir (#,' ' ou ?).

bombasExplodem :: Mapa -> [Coordenadas]

A função bombasExplodem diz as coordenadas das bombas que vão explodir.

bombasExplodemAux :: [Coordenadas] -> [Int] -> [Coordenadas]

A função bombasExplodemAux é a principal função auxiliar da função bombasExplodem.

raiosExplodem :: Mapa -> [[Raio]]

A função raiosExplodem diz as listas dos raios das bombas que vão explodir.

Exemplo de utilização:

>>> raiosExplodem m = [[1,2],[1,2,3]]

(m= ["######## ?",""," ? ? ? ","","#########","* 4 5 0 2 1","* 5 6 1 3 1","* 7 6 0 1 8","0 5 4","1 6 5"])

raiosExplodemAux :: [[Raio]] -> [Int] -> [[Raio]]

A função raiosExplodemAux é a principal função auxiliar da função raiosExplodem.

(!-!) :: [a] -> Int -> a

A função (!-!) é uma adaptação da função !!, já existente na Prelude do Haskell.

quaisExplodem :: Mapa -> [Int]

A função quaisExplodem dá a posição das bombas que vão explodir (relativamente a info nova das bombas).

quaisExplodem1 :: Mapa -> [Int]

A função quaisExplodem1 é a principal função auxiliar da função quaisExplodem.

listasraios :: Mapa -> [[Raio]]

A função listasraios diz as listas dos raios de todas bombas presentes no mapa.

Exemplo de utilização:

>>> listasraios m = [[1,2],[1,2,3],[1]]

(m= ["######## ?",""," ? ? ? ","","#########","* 4 5 0 2 1","* 5 6 1 3 1","* 7 6 0 1 8","0 5 4","1 6 5"])

listasraiosAux :: [Raio] -> [[Raio]]

A função listasraiosAux é a principal função auxiliar da função listasraios.

raioBombaInt :: Mapa -> [Raio]

A função raioBombaInt dá o raio de cada bomba plantada em [Int].

raioBomba :: Mapa -> [String]

A função raioBomba dá o raio de cada bomba plantada em [String].

raioBombaAux :: [[String]] -> [String]

novoMapaSE :: Mapa -> Mapa

A função novoMapaSE retira 1 instante de tempo a todas as bombas presentes no mapa (após as que tinham que explodir efetivamente explodirem).

colocaLinhas :: Mapa -> [String] -> Mapa

A função colocaLinhas coloca as linhas com as informações novas das bombas (nomeadamente os instantes de tempo), no mapa.

eliminaLinhas :: Mapa -> Mapa

A função eliminaLinhas elimina do mapa as linhas relativas às informações (antigas) das bombas.

novaInfoBombas :: Mapa -> [String]

A função novaInfoBombas põe a info (nova) das bombas no formato desejado ("unwords").

unwords2 :: [[String]] -> [String]

newTime :: Mapa -> [[String]]

A função newTime substitui instantes de tempo antigos pelos novos.

newTimeAux :: [[String]] -> [String] -> [[String]]

A função newTimeAux é a principal função auxiliar da função nemTime.

substituiu :: [String] -> String -> [String]

words1 :: Mapa -> [[String]]

A função words1 põe a info (antiga) das bombas em formato "words".

words2 :: [String] -> [[String]]

vaiExplodir :: Mapa -> Bool

A função vaiExplodir averigua se alguma das bombas plantadas vai explodir no novo instante de tempo.

vaiExplodirAux :: [Int] -> Bool

A função vaiExplodirAux é a principal função auxiliar da função vaiExplodir.

novoIString :: Mapa -> [String]

A função novoIString indica quantos instantes faltam para as bombas explodirem em formato String.

novoIStringAux :: [Int] -> [String]

A função novoIStringAux é a principal função auxiliar da função novoIString.

novoInst :: Mapa -> [Int]

A função novoInst indica quantos instantes faltam para as bombas explodirem no novo instante de tempo.

novoInstAux :: [Int] -> [Int]

A função novoInstAux é a principal função auxiliar da função novoInst.

qtsInstantes :: Mapa -> [Int]

A função qtsInstantes indica quantos instantes faltam para as bombas explodirem em formato [Int].

string2Int :: [String] -> [Int]

qtsInstantes1 :: Mapa -> [String]

A função qtsInstantes1 indica quantos instantes faltam para as bombas explodirem em formato [String].

qtsInstantes1Aux :: [String] -> [String]

infoBombas :: Mapa -> [String]

A função infoBombas pega nas linhas das informações das bombas.

listapu :: Mapa -> [Coordenadas]

A função lista pu dá como resposta as coordenadas dos power ups presentes no mapa.

coordenadasPU1 :: [String] -> [Coordenadas]

coordenadasPU2 :: [String] -> [String]

tirarPUeCV :: [String] -> [String]

auxTirar :: String -> String

powerUp :: [String] -> [String]

coord :: String -> Coordenadas

A função coord devolve as coordenadas no tipo (Int,Int).

coord1 :: String -> String

coord2 :: String -> String

listajog :: Mapa -> [Coordenadas]

A função listajog dá como resultado a lista das coordenadas dos jogadores no mapa.

toInt :: [[String]] -> [Coordenadas]

toInt1 :: [String] -> Coordenadas

listajog1 :: Mapa -> [[String]]

listajog2 :: [[String]] -> [[String]]

listajog3 :: [String] -> [String]

tail1 :: Mapa -> [[String]]

tail2 :: [[String]] -> [[String]]

words3 :: Mapa -> [[String]]

linhaplayers :: Mapa -> [String]

A função linhaplayers dá como resultado as linhas alusivas às informaçoes dos jogadores.

listabomb :: Mapa -> [Coordenadas]

A função listabomb dá como resultado a lista das coordenadas das bombas plantadas no mapa.

listabomb1 :: Mapa -> [[String]]

tail3 :: Mapa -> [[String]]

words4 :: Mapa -> [[String]]

novomapaSemBombaExpl :: Mapa -> Coordenadas -> Mapa

A função novomapaSemBombaExpl dá como resultado o mapa sem a linha relativa à bomba explodida.

novomapaSemBombaExplAux :: Mapa -> Mapa -> Coordenadas -> Mapa

bombexistentes :: Mapa -> Coordenadas -> [String]

A função bombexistentes dá como resultado as info das bombas que se mantêm no mapa (ainda não explodiram).

bombexistentes1 :: [String] -> Int -> [String]

qualbomb :: Mapa -> Coordenadas -> Int

A função qualbomb, relativamente às infos das bombas, diz qual é a posição da info que precisa de ser apagada.

eliminalinhasbomb :: Mapa -> Mapa

A função eliminalinhasbomb elimina todas as linhas das info das bombas colocadas.

novomapaSemPI :: Mapa -> Mapa -> Coordenadas -> Coordenadas -> Mapa

A função novomapaSemPI dá como resultado o mapa sem o tijolo que foi atingido pelas chamas.

novaLinha :: Mapa -> Coordenadas -> String

A função novaLinha dá como resultado a nova linha atualizada (sem o ?).

tiraPI :: String -> Coordenadas -> String

A função tiraPI substitui o ? por ' '.

pegaLinha :: Mapa -> Coordenadas -> String

A função pegaLinha pega na linha onde está o ? que procuramos.

novomapaComInst1 :: Mapa -> Coordenadas -> Mapa

A função novomapaComInst1 dá como resultado o mapa com a bomba que foi atingida por uma bomba que explodiu, com os instantes de tempo devidamente alterados (neste caso, com instantes = 2, porque a função novoMapaSE vai retirar mais 1 instante de tempo a cada bomba).

novomapaComInst1Aux :: Mapa -> Mapa -> Coordenadas -> Mapa

alterainstb :: Mapa -> Coordenadas -> [String]

A função alterainstb altera informação da bomba atingida pela chama (instante fica = 2).

alterainstb1 :: Mapa -> Int -> [String]

alterainstante :: String -> String

alterainstante1 :: String -> [String]

alterainstanteAux :: String -> [String]

qualbombating :: Mapa -> Coordenadas -> Int

A função qualbombating, relativamente as infos das bombas, diz qual é a posição da info que precisa de mudar instante para 2.

novomapaSemPU :: Mapa -> Coordenadas -> Mapa

A função novomapaSemPU elimina a linha do power up que foi atingido pela chama de uma bomba.

novomapaSemPUAux :: Mapa -> Mapa -> Coordenadas -> Mapa

puexistentes :: Mapa -> Coordenadas -> [String]

A função alterainstb devolve somente a informação das linhas do power ups que se mantêm no mapa.

puexistentes1 :: [String] -> Int -> [String]

qualpu1 :: [String] -> Coordenadas -> Int

A função qualpu1, relativamente as infos dos power ups, diz qual é a posicao da info que precisa de ser apagada.

infoPU :: Mapa -> [String]

A função infoPU devolve as informações dos jogadores que se encontram em jogo.

eliminalinhaspu :: Mapa -> Mapa

A função eliminalinhaspu elimina do mapa as linhas relativas às informações dos power ups.

novomapaSemJog :: Mapa -> Coordenadas -> Mapa

A função novomapaSemJog elimina do mapa as linhas relativas aos jogadores que perderam.

jogvivos :: Mapa -> Coordenadas -> [String]

A função jogvivos devolve as linhas dos jogadores que estão vivos.

jogvivos1 :: Mapa -> [Int] -> [String]

funcaoquesubtrai1 :: [Int] -> [Int]

qualjogador1 :: Mapa -> Coordenadas -> [Int]

A função qualjogador1, relativamente as infos dos jogadores, diz qual é a posição da info que precisa de ser apagada.

my_elemIndices :: Eq a => a -> [a] -> [Int]

infoJogadores :: Mapa -> [String]

A função infoJogadores devolve as informações dos jogadores no mapa.

eliminalinhasjog :: Mapa -> Mapa

A função eliminalinhasjog devolve o mapa sem as linhas relativas às informações dos jogadores.