Tarefa6_li1g172

Documentation

type Coordenadas = (Int, Int)

type Mapa = [String]

type NBot = Int

type TimeLeft = Int

type Raio = Int

type DimensaoMapa = Int

bot :: Mapa -> NBot -> TimeLeft -> Maybe Char

Description : Módulo Haskell que contém a função que implementa uma estratégia de combate (função bot).

Este módulo contém funções recursivas e não recursivas, com o intuito de devolver o comando do bot que joga bomberman automaticamente.

coordBot :: Mapa -> NBot -> Coordenadas

A função coordBot dá como resposta as coordenadas do mapa em que o bot se encontra.

coordEmString :: Mapa -> NBot -> [String]

linhainfoBot :: Mapa -> NBot -> String

A função linhainfoBot dá como resposta a linha da informação do bot.

coordAdversarios :: Mapa -> NBot -> [Coordenadas]

A função coordAdversarios dá como resposta as coordenadas dos outros jogadores (adversários).

toInt :: [[String]] -> [Coordenadas]

listajog :: Mapa -> NBot -> [[String]]

listajog1 :: [[String]] -> [[String]]

tail1 :: Mapa -> NBot -> [[String]]

words1 :: Mapa -> NBot -> [[String]]

linhasinfoAdversarios :: Mapa -> NBot -> [String]

A função linhainfoAdversarios dá como resposta as linhas das informações dos adversários.

linhasinfoAdvAux :: [String] -> [String]

A função linhasinfoAdvAux é a principal função auxiliar da função linhasinfoAdversarios.

eliminainfoBot :: Mapa -> NBot -> Mapa

A função eliminainfoBot elimina do mapa a linha respetiva à info do bot.

dimensaomapa :: Mapa -> Int

A função dimensaomapa dá como resposta a dimensão do mapa do estado de jogo em questão.

listabomb :: Mapa -> [Coordenadas]

A função listabomb diz as coordenadas das bombas que estão plantadas.

listabomb1 :: Mapa -> [[String]]

tail2 :: Mapa -> [[String]]

words2 :: Mapa -> [[String]]

infoBombas :: Mapa -> [String]

A função infoBombas pega nas linhas relativas às informações das bombas.

raioBomba :: Mapa -> [Raio]

A função raioBomba dá como resposta os raios das bombas plantadas.

raioBomba1 :: Mapa -> [String]

A função raioBomba1 dá como resposta os raios das bombas plantadas em [String].

instantesBomba :: Mapa -> [TimeLeft]

A função instantesBomba diz os instantes que faltam para as bombas plantadas explodirem.

instantesBomba1 :: Mapa -> [String]

A função instantesBomba1 é a principal função auxiliar da função instantesBomba.

listapudestapados :: Mapa -> [Coordenadas]

A função listapudestapados dá a lista dos powerups que se encontram destapados (tijolo que o encobria já foi explodido).

down :: Mapa -> NBot -> Char

A função down diz o que está abaixo do bot (#,' ' ou ?).

up :: Mapa -> NBot -> Char

A função up diz o que está acima do bot (#,' ' ou ?).

left :: Mapa -> NBot -> Char

A função left diz o que está à esquerda do bot (#,' ' ou ?).

right :: Mapa -> NBot -> Char

A função right diz o que está à direita do bot (#,' ' ou ?).

downc :: Mapa -> NBot -> Coordenadas

A função downc diz qual é a coordenada que está imediatamente abaixo da posição do bot.

downcAux :: Coordenadas -> Coordenadas

A função downcAux diz qual é a coordenada que está imediatamente abaixo de uma determinada coordenada.

upc :: Mapa -> NBot -> Coordenadas

A função upc diz qual é a coordenada que está imediatamente acima da posição do bot.

upcAux :: Coordenadas -> Coordenadas

A função upcAux diz qual é a coordenada que está imediatamente acima de uma determinada coordenada.

leftc :: Mapa -> NBot -> Coordenadas

A função leftc diz qual é a coordenada que está imediatamente à esquerda da posição do bot.

leftcAux :: Coordenadas -> Coordenadas

A função leftcAux diz qual é a coordenada que está imediatamente à esquerda de uma determinada coordenada.

rightc :: Mapa -> NBot -> Coordenadas

A função rightc diz qual é a coordenada que está imediatamente à direita da posição do bot.

rightcAux :: Coordenadas -> Coordenadas

A função rightcAux diz qual é a coordenada que está imediatamente à direita de uma determinada coordenada.

oqueestaD :: Mapa -> Coordenadas -> Raio -> Char

A função oqueestaD diz o que está x células abaixo de uma determinada coordenada.

oqueestaU :: Mapa -> Coordenadas -> Raio -> Char

A função oqueestaU diz o que está x células acima de uma determinada coordenada.

oqueestaL :: Mapa -> Coordenadas -> Raio -> Char

A função oqueestaL diz o que está x células à esquerda de uma determinada coordenada.

oqueestaR :: Mapa -> Coordenadas -> Raio -> Char

A função oqueestaR diz o que está x células à direita de uma determinada coordenada.

downpormenor :: (Int, Int) -> Raio -> [String] -> String

A função downpormenor vê com mais pormenor, o que está x coordenadas abaixo de uma determinada coordenada.

Respostas possíveis:

Pedra;

Tijolo;

Powerup;

Bomb;

Jogador;

Nada.

espacoVazioD :: (Int, Int) -> Raio -> [String] -> String

A função espacoVazioD é a principal função auxiliar da função downpormenor.

uppormenor :: (Int, Int) -> Raio -> [String] -> String

A função uppormenor vê com mais pormenor, o que está x coordenadas acima de uma determinada coordenada.

Respostas possíveis:

Pedra;

Tijolo;

Powerup;

Bomb;

Jogador;

Nada.

espacoVazioU :: (Int, Int) -> Raio -> [String] -> String

A função espacoVazioU é a principal função auxiliar da função uppormenor.

leftpormenor :: (Int, Int) -> Raio -> [String] -> String

A função leftpormenor vê com mais pormenor, o que está x coordenadas à esquerda de uma determinada coordenada.

Respostas possíveis:

Pedra;

Tijolo;

Powerup;

Bomb;

Jogador;

Nada.

espacoVazioL :: (Int, Int) -> Raio -> [String] -> String

A função espacoVazioL é a principal função auxiliar da função leftpormenor.

rightpormenor :: (Int, Int) -> Raio -> [String] -> String

A função rightpormenor vê com mais pormenor, o que está x coordenadas à direita de uma determinada coordenada.

Respostas possíveis:

Pedra;

Tijolo;

Powerup;

Bomb;

Jogador;

Nada.

espacoVazioR :: (Int, Int) -> Raio -> [String] -> String

A função espacoVazioR é a principal função auxiliar da função rightpormenor.

chamasBombas :: Mapa -> [Coordenadas]

A função chamasBombas dá como resultado a lista das coordenadas que as chamas das bombas atingem.

Tendo em consideração que as chamas não atravessam:

As pedras;

Os tijolos;

Os espaços vazios com power up.

chamasBombasAux :: Mapa -> [Coordenadas] -> [[Int]] -> [Coordenadas]

A função chamasBombasAux é a principal função auxiliar da função chamasBombas.

Esta função concatena o resultado das 4 primeiras (por vezes únicas, depende do raio de destruição da bomba) direções (cima, baixo, esquerda e direita) que as chamas percorrem; e tem como passo recursivo, percorrer as seguintes 4 direções.

eliminaChamasD :: Mapa -> Coordenadas -> [Raio] -> [Coordenadas]

A função eliminaChamasD, tem como parâmetros:

O mapa;

A coordenada de uma bomba que explode;

A lista do raio que essa bomba alcança (de 1 até ao raio da bomba).

Esta função pára de contabilizar as coordenadas das chamas das bombas a partir do momento em que encontra uma pedra, um tijolo, ou um espaço vazio com power up.

Dá como resultado as coordenadas que as chamas das bombas atingem, neste caso inferiores (para baixo) do local da bomba.

eliminaChamasU :: Mapa -> Coordenadas -> [Raio] -> [Coordenadas]

A função eliminaChamasU, tem como parâmetros:

O mapa;

A coordenada de uma bomba que explode;

A lista do raio que essa bomba alcança (de 1 até ao raio da bomba).

Esta função pára de contabilizar as coordenadas das chamas das bombas a partir do momento em que encontra uma pedra, um tijolo, ou um espaço vazio com power up.

Dá como resultado as coordenadas que as chamas das bombas atingem, neste caso superiores (para cima) do local da bomba.

eliminaChamasL :: Mapa -> Coordenadas -> [Raio] -> [Coordenadas]

A função eliminaChamasL, tem como parâmetros:

O mapa;

A coordenada de uma bomba que explode;

A lista do raio que essa bomba alcança (de 1 até ao raio da bomba).

Esta função pára de contabilizar as coordenadas das chamas das bombas a partir do momento em que encontra uma pedra, um tijolo, ou um espaço vazio com power up.

Dá como resultado as coordenadas que as chamas das bombas atingem, neste caso para a esquerda do local da bomba.

eliminaChamasR :: Mapa -> Coordenadas -> [Raio] -> [Coordenadas]

A função eliminaChamasR, tem como parâmetros:

O mapa;

A coordenada de uma bomba que explode;

A lista do raio que essa bomba alcança (de 1 até ao raio da bomba).

Esta função pára de contabilizar as coordenadas das chamas das bombas a partir do momento em que encontra uma pedra, um tijolo, ou um espaço vazio com power up.

Dá como resultado as coordenadas que as chamas das bombas atingem, neste caso __para a direita do local da bomba.

listaRaios :: Mapa -> [[Int]]

A função listaRaios diz as listas dos raios das bombas plantadas.

tiraCoordPedras :: Mapa -> [Coordenadas] -> [Coordenadas]

A função tiraCoordPedras, tem como parâmetros o mapa e a lista das coordenadas que as chamas das bombas atingem, e retira dessa lista as coordenadas que correspondem a pedras.

eliminaRepetidos :: [Coordenadas] -> [Coordenadas]

Na função eliminaRepetidos, dada uma lista de coordenadas, elimina dessa lista as coordenadas que se encontram repetidas. (Isso acontece quando bombas diferentes explodem e atingem as mesmas células).

eliminaCoord :: Coordenadas -> [Coordenadas] -> [Coordenadas]

A função eliminaCoord é a função auxiliar da função eliminaRepetidos.

foraDePerigoEm :: [String] -> [Coordenadas]

A função foraDePerigoEm calcula as coordenadas em que o bot está fora de perigo.

apagaCoord :: [Coordenadas] -> [Coordenadas] -> [Coordenadas]

A função apagaCoord recebe como parâmetros 2 listas de coordenadas, e elimina da 2ª, as coordenadas da 1ª.

retira :: Coordenadas -> [Coordenadas] -> [Coordenadas]

A função retira é a função auxiliar da função apagaCoord.

botEmPerigo :: Mapa -> NBot -> Bool

A função botEmPerigo averigua se na coordenada em que o bot se encontra, ele está em perigo ou não (se alguma chama atinge essa coordenada).

botEmPerigoD :: Mapa -> NBot -> Bool

A função botEmPerigoD averigua se na coordenada abaixo do bot, ele ficará em perigo ou não (se alguma chama atinge essa coordenada).

botEmPerigoU :: Mapa -> NBot -> Bool

A função botEmPerigoT averigua se na coordenada acima do bot, ele ficará em perigo ou não (se alguma chama atinge essa coordenada).

botEmPerigoL :: Mapa -> NBot -> Bool

A função botEmPerigoL averigua se na coordenada à esquerda do bot, ele ficará em perigo ou não (se alguma chama atinge essa coordenada).

botEmPerigoR :: Mapa -> NBot -> Bool

A função botEmPerigoR averigua se na coordenada à direita do bot, ele ficará em perigo ou não (se alguma chama atinge essa coordenada).

analisaParede :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaParede, consoante o bot esteja rodeado por parede (pedra ou tijolo), excluí essas direções, e põe o bot a analisar somente as direções para as quais ele pode efetivamente ir.

analisaU :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaU analisa conveniente o que o bot deve fazer quando só tem a direção "up" disponível.

analisaL :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaL analisa conveniente o que o bot deve fazer quando só tem a direção "left" disponível.

analisaR :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaR analisa conveniente o que o bot deve fazer quando só tem a direção "right" disponível.

analisaD :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaD analisa conveniente o que o bot deve fazer quando só tem a direção "down" disponível.

analisaUL :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaUL analisa conveniente o que o bot deve fazer quando só tem as direções "up" e "left" disponíveis.

analisaUR :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaUR analisa conveniente o que o bot deve fazer quando só tem as direções "up" e "right" disponíveis.

analisaLR :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaLR analisa conveniente o que o bot deve fazer quando só tem as direções "left" e "right" disponíveis.

analisaDU :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaDU analisa conveniente o que o bot deve fazer quando só tem as direções "down" e "up" disponíveis.

analisaLD :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaLD analisa conveniente o que o bot deve fazer quando só tem as direções "left" e "down" disponíveis.

analisaRD :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaRD analisa conveniente o que o bot deve fazer quando só tem as direções "right" e "down" disponíveis.

analisaULR :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaULR analisa conveniente o que o bot deve fazer quando só tem as direções "up", "left" e "right" disponíveis.

analisaDUR :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaDUR analisa conveniente o que o bot deve fazer quando só tem as direções "down", "up" e "right" disponíveis.

analisaDUL :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaDUL analisa conveniente o que o bot deve fazer quando só tem as direções "down", "up" e "left" disponíveis.

analisaDLR :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaDLR analisa conveniente o que o bot deve fazer quando só tem as direções "down", "left" e "right" disponíveis.

analisaDULR :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função analisaDULR analisa conveniente o que o bot deve fazer quando tem todas as direções disponíveis.

tempoATerminarDULR :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarDULR averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralDULR :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralDULR encaminha o bot para a coordenada mais próxima da coordenada objetivo.

tempoATerminarDLR :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarDLR averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralDLR :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralDLR encaminha o bot para a coordenada mais próxima da coordenada objetivo.

tempoATerminarDUL :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarDUL averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralDUL :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralDUL encaminha o bot para a coordenada mais próxima da coordenada objetivo.

tempoATerminarDUR :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarDUR averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralDUR :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralDUR encaminha o bot para a coordenada mais próxima da coordenada objetivo.

tempoATerminarULR :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarULR averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralULR :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralULR encaminha o bot para a coordenada mais próxima da coordenada objetivo.

tempoATerminarD :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarD averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralD :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralD encaminha o bot para a coordenada mais próxima da coordenada objetivo.

tempoATerminarU :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarU averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralU :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralU encaminha o bot para a coordenada mais próxima da coordenada objetivo.

tempoATerminarR :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarR averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralR :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralR encaminha o bot para a coordenada mais próxima da coordenada objetivo.

tempoATerminarL :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarL averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralL :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralL encaminha o bot para a coordenada mais próxima da coordenada objetivo.

tempoATerminarRD :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarRD averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralRD :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralRD encaminha o bot para a coordenada mais próxima da coordenada objetivo.

tempoATerminarLD :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarLD averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralLD :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralLD encaminha o bot para a coordenada mais próxima da coordenada objetivo.

tempoATerminarDU :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarDU averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralDU :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralDU encaminha o bot para a coordenada mais próxima da coordenada objetivo.

tempoATerminarLR :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarLR averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralLR :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralLR encaminha o bot para a coordenada mais próxima da coordenada objetivo.

tempoATerminarUR :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarUR averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralUR :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralUR encaminha o bot para a coordenada mais próxima da coordenada objetivo.

tempoATerminarUL :: Mapa -> TimeLeft -> NBot -> Maybe Char

A função tempoATerminarUL averigua se faltam 10 ou menos instantes de tempo para a espiral começar.

Caso afirmativo, o bot procura fugir da espiral (ir para o centro do mapa);

Caso contrário, o bot vai tentar plantar bombas (se vantajoso).

fugirEspiralUL :: Coordenadas -> Coordenadas -> Maybe Char

A função fugirEspiralUL encaminha o bot para a coordenada mais próxima da coordenada objetivo.

coordenadaObjetivo :: Mapa -> Coordenadas

A função coordenadaObjetivo averigua se o mapa em questão tem como coordenada central uma pedra.

Caso afirmativo, a coordenada objetivo é a coordenada ao lado esquerdo da coordenada central (neste caso, última célula onde cai pedra);

Caso contrário, a coordenada objetivo é a coordenada central do mapa.

coordenadaObjetivo1 :: Mapa -> Coordenadas

A função coordenadaObjetivo1 dá como resposta a coordenada ao lado esquerdo da coordenada central do mapa.

coordenadaObjetivo2 :: Mapa -> Coordenadas

A função coordenadaObjetivo2 dá como resposta a coordenada central do mapa.

listasDimensoesCoordenadaCentroP :: Int -> [Int]

A função listasDimensoesCoordenadaCentroP dá a lista de alguns mapas que possuem uma pedra na coordenada central do mapa.

dist :: Coordenadas -> Coordenadas -> Float

A função dist dá como resultado a distância entre 2 coordenadas.

plantaBombasDULR :: Mapa -> NBot -> Maybe Char

A função plantaBombasDULR faz o bot plantar uma bomba se tiver se o mesmo estiver rodeado por algum tijolo ou jogador; caso não esteja, o bot irá procurar algum power up destapado, tijolo ou jogador (se estiverem em coordenadas próximas).

procuraPowerDULR :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerDULR faz com que o bot averigue se tem algum power up próximo de si (num raio de 3 coordenadas à sua volta).

procuraPowerDULRAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerDULRAux é a principal função auxiliar da função procuraPowerDULR.

procuraTijolosDULR :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosDULR procura tijolos no mapa (num raio de 3 coordenadas), e em caso afirmativo o bot vai nessa direção.

procuraTijolosDULRAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosDULRAux é a principal função auxiliar da função procuraTijolosDULR.

procuraJogadoresDULR :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresDULR procura jogadores no mapa, num raio de 3 coordenadas.

Caso não existam, vai simplesmente para baixo.

procuraJogadoresDULRAux :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresDULRAux é a principal função auxiliar da função procuraJogadoresDULR.

plantaBombasDLR :: Mapa -> NBot -> Maybe Char

A função plantaBombasDLR faz o bot plantar uma bomba se tiver se o mesmo estiver rodeado por algum tijolo ou jogador; caso não esteja, o bot irá procurar algum power up destapado, tijolo ou jogador (se estiverem em coordenadas próximas).

procuraPowerDLR :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerDLR faz com que o bot averigue se tem algum power up próximo de si (num raio de 3 coordenadas à sua volta).

procuraPowerDLRAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerDLRAux é a principal função auxiliar da função procuraPowerDLR.

procuraTijolosDLR :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosDLR procura tijolos no mapa (num raio de 3 coordenadas), e em caso afirmativo o bot vai nessa direção.

procuraTijolosDLRAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosDLRAux é a principal função auxiliar da função procuraTijolosDLR.

procuraJogadoresDLR :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresDLR procura jogadores no mapa, num raio de 3 coordenadas.

Caso não existam, vai simplesmente para baixo.

procuraJogadoresDLRAux :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresDLRAux é a principal função auxiliar da função procuraJogadoresDLR.

plantaBombasDUL :: Mapa -> NBot -> Maybe Char

A função plantaBombasDUL faz o bot plantar uma bomba se tiver se o mesmo estiver rodeado por algum tijolo ou jogador; caso não esteja, o bot irá procurar algum power up destapado, tijolo ou jogador (se estiverem em coordenadas próximas).

procuraPowerDUL :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerDUL faz com que o bot averigue se tem algum power up próximo de si (num raio de 3 coordenadas à sua volta).

procuraPowerDULAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerDULAux é a principal função auxiliar da função procuraPowerDUL.

procuraTijolosDUL :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosDUL procura tijolos no mapa (num raio de 3 coordenadas), e em caso afirmativo o bot vai nessa direção.

procuraTijolosDULAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosDULAux é a principal função auxiliar da função procuraTijolosDUL.

procuraJogadoresDUL :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresDUL procura jogadores no mapa, num raio de 3 coordenadas.

Caso não existam, vai simplesmente para baixo.

procuraJogadoresDULAux :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresDULAux é a principal função auxiliar da função procuraJogadoresDUL.

plantaBombasDUR :: Mapa -> NBot -> Maybe Char

A função plantaBombasDUR faz o bot plantar uma bomba se tiver se o mesmo estiver rodeado por algum tijolo ou jogador; caso não esteja, o bot irá procurar algum power up destapado, tijolo ou jogador (se estiverem em coordenadas próximas).

procuraPowerDUR :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerDUR faz com que o bot averigue se tem algum power up próximo de si (num raio de 3 coordenadas à sua volta).

procuraPowerDURAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerDURAux é a principal função auxiliar da função procuraPowerDUR.

procuraTijolosDUR :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosDUR procura tijolos no mapa (num raio de 3 coordenadas), e em caso afirmativo o bot vai nessa direção.

procuraTijolosDURAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosDURAux é a principal função auxiliar da função procuraTijolosDUR.

procuraJogadoresDUR :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresDUR procura jogadores no mapa, num raio de 3 coordenadas.

Caso não existam, vai simplesmente para baixo.

procuraJogadoresDURAux :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresDURAux é a principal função auxiliar da função procuraJogadoresDUR.

plantaBombasULR :: Mapa -> NBot -> Maybe Char

A função plantaBombasULR faz o bot plantar uma bomba se tiver se o mesmo estiver rodeado por algum tijolo ou jogador; caso não esteja, o bot irá procurar algum power up destapado, tijolo ou jogador (se estiverem em coordenadas próximas).

procuraPowerULR :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerULR faz com que o bot averigue se tem algum power up próximo de si (num raio de 3 coordenadas à sua volta).

procuraPowerULRAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerULRAux é a principal função auxiliar da função procuraPowerULR.

procuraTijolosULR :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosULR procura tijolos no mapa (num raio de 3 coordenadas), e em caso afirmativo o bot vai nessa direção.

procuraTijolosULRAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosULRAux é a principal função auxiliar da função procuraTijolosULR.

procuraJogadoresULR :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresULR procura jogadores no mapa, num raio de 3 coordenadas.

Caso não existam, vai simplesmente para a direita.

procuraJogadoresULRAux :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresULRAux é a principal função auxiliar da função procuraJogadoresULR.

plantaBombasD :: Mapa -> NBot -> Maybe Char

plantaBombasU :: Mapa -> NBot -> Maybe Char

plantaBombasL :: Mapa -> NBot -> Maybe Char

plantaBombasR :: Mapa -> NBot -> Maybe Char

plantaBombasRD :: Mapa -> NBot -> Maybe Char

A função plantaBombasRD faz o bot plantar uma bomba se tiver se o mesmo estiver rodeado por algum tijolo ou jogador; caso não esteja, o bot irá procurar algum power up destapado, tijolo ou jogador (se estiverem em coordenadas próximas).

procuraPowerRD :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerRD faz com que o bot averigue se tem algum power up próximo de si (num raio de 3 coordenadas à sua volta).

procuraPowerRDAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerRDAux é a principal função auxiliar da função procuraPowerRD.

procuraTijolosRD :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosRD procura tijolos no mapa (num raio de 3 coordenadas), e em caso afirmativo o bot vai nessa direção.

procuraTijolosRDAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosRDAux é a principal função auxiliar da função procuraTijolosRD.

procuraJogadoresRD :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresRD procura jogadores no mapa, num raio de 3 coordenadas.

Caso não existam, vai simplesmente para baixo.

procuraJogadoresRDAux :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresRDAux é a principal função auxiliar da função procuraJogadoresRD.

plantaBombasLD :: Mapa -> NBot -> Maybe Char

A função plantaBombasLD faz o bot plantar uma bomba se tiver se o mesmo estiver rodeado por algum tijolo ou jogador; caso não esteja, o bot irá procurar algum power up destapado, tijolo ou jogador (se estiverem em coordenadas próximas).

procuraPowerLD :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerLD faz com que o bot averigue se tem algum power up próximo de si (num raio de 3 coordenadas à sua volta).

procuraPowerLDAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerLDAux é a principal função auxiliar da função procuraPowerLD.

procuraTijolosLD :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosLD procura tijolos no mapa (num raio de 3 coordenadas), e em caso afirmativo o bot vai nessa direção.

procuraTijolosLDAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosLDAux é a principal função auxiliar da função procuraTijolosLD.

procuraJogadoresLD :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresLD procura jogadores no mapa, num raio de 3 coordenadas.

Caso não existam, vai simplesmente para baixo.

procuraJogadoresLDAux :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresLDAux é a principal função auxiliar da função procuraJogadoresLD.

plantaBombasDU :: Mapa -> NBot -> Maybe Char

A função plantaBombasDU faz o bot plantar uma bomba se tiver se o mesmo estiver rodeado por algum tijolo ou jogador; caso não esteja, o bot irá procurar algum power up destapado, tijolo ou jogador (se estiverem em coordenadas próximas).

procuraPowerDU :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerDU faz com que o bot averigue se tem algum power up próximo de si (num raio de 3 coordenadas à sua volta).

procuraPowerDUAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerDUAux é a principal função auxiliar da função procuraPowerDU.

procuraTijolosDU :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosDU procura tijolos no mapa (num raio de 3 coordenadas), e em caso afirmativo o bot vai nessa direção.

procuraTijolosDUAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosDUAux é a principal função auxiliar da função procuraTijolosDU.

procuraJogadoresDU :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresDU procura jogadores no mapa, num raio de 3 coordenadas.

Caso não existam, vai simplesmente para baixo.

procuraJogadoresDUAux :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresDUAux é a principal função auxiliar da função procuraJogadoresDU.

plantaBombasLR :: Mapa -> NBot -> Maybe Char

A função plantaBombasLR faz o bot plantar uma bomba se tiver se o mesmo estiver rodeado por algum tijolo ou jogador; caso não esteja, o bot irá procurar algum power up destapado, tijolo ou jogador (se estiverem em coordenadas próximas).

procuraPowerLR :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerLR faz com que o bot averigue se tem algum power up próximo de si (num raio de 3 coordenadas à sua volta).

procuraPowerLRAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerLRAux é a principal função auxiliar da função procuraPowerLR.

procuraTijolosLR :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosLR procura tijolos no mapa (num raio de 3 coordenadas), e em caso afirmativo o bot vai nessa direção.

procuraTijolosLRAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosLRAux é a principal função auxiliar da função procuraTijolosLR.

procuraJogadoresLR :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresLR procura jogadores no mapa, num raio de 3 coordenadas.

Caso não existam, vai simplesmente para a direita.

procuraJogadoresLRAux :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresLRAux é a principal função auxiliar da função procuraJogadoresLR.

plantaBombasUR :: Mapa -> NBot -> Maybe Char

A função plantaBombasUR faz o bot plantar uma bomba se tiver se o mesmo estiver rodeado por algum tijolo ou jogador; caso não esteja, o bot irá procurar algum power up destapado, tijolo ou jogador (se estiverem em coordenadas próximas).

procuraPowerUR :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerUR faz com que o bot averigue se tem algum power up próximo de si (num raio de 3 coordenadas à sua volta).

procuraPowerURAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerURAux é a principal função auxiliar da função procuraPowerUR.

procuraTijolosUR :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosUR procura tijolos no mapa (num raio de 3 coordenadas), e em caso afirmativo o bot vai nessa direção.

procuraTijolosURAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosURAux é a principal função auxiliar da função procuraTijolosUR.

procuraJogadoresUR :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresUR procura jogadores no mapa, num raio de 3 coordenadas.

Caso não existam, vai simplesmente para a direita.

procuraJogadoresURAux :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresURAux é a principal função auxiliar da função procuraJogadoresUR.

plantaBombasUL :: Mapa -> NBot -> Maybe Char

A função plantaBombasUL faz o bot plantar uma bomba se tiver se o mesmo estiver rodeado por algum tijolo ou jogador; caso não esteja, o bot irá procurar algum power up destapado, tijolo ou jogador (se estiverem em coordenadas próximas).

procuraPowerUL :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerUL faz com que o bot averigue se tem algum power up próximo de si (num raio de 3 coordenadas à sua volta).

procuraPowerULAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Int -> Maybe Char

A função procuraPowerULAux é a principal função auxiliar da função procuraPowerUL.

procuraTijolosUL :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosUL procura tijolos no mapa (num raio de 3 coordenadas), e em caso afirmativo o bot vai nessa direção.

procuraTijolosULAux :: Mapa -> NBot -> Coordenadas -> Int -> Int -> Maybe Char

A função procuraTijolosULAux é a principal função auxiliar da função procuraTijolosUL.

procuraJogadoresUL :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresUL procura jogadores no mapa, num raio de 3 coordenadas.

Caso não existam, o bot não faz anda.

procuraJogadoresULAux :: Mapa -> NBot -> Coordenadas -> Int -> Maybe Char

A função procuraJogadoresULAux é a principal função auxiliar da função procuraJogadoresUL.

temTijolosAVolta :: Mapa -> NBot -> Bool

A função temTijolosAVolta vê se o bot está "cercado" por algum tijolo.

temTijolosAVoltaAux :: Mapa -> Coordenadas -> Bool

A função temTijolosAVoltaAux é a principal função auxiliar da função temTijolosAVolta.

temJogadoresAVolta :: Mapa -> NBot -> Bool

A função temJogadoresAVolta vê se o bot tem jogadores à sua volta (ou na própria posição).

fugirDULR :: Mapa -> NBot -> Maybe Char

A função fugirDULR ajuda o bot a fugir das bombas, quando ele está indeciso entre todas as direções.

fugirDULRAux :: Mapa -> [Coordenadas] -> Maybe Char

estudaDULR1 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDULR2 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDULR3 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDULR4 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDULR11 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDULR22 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDULR33 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDULR44 :: Mapa -> Coordenadas -> Maybe Char

fugirDLR :: Mapa -> NBot -> Maybe Char

A função fugirDLR ajuda o bot a fugir das bombas, quando ele está indeciso entre down, left e right.

fugirDLRAux :: Mapa -> [Coordenadas] -> Maybe Char

estudaDLR1 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDLR2 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDLR3 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDLR11 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDLR22 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDLR33 :: Mapa -> Coordenadas -> Maybe Char

fugirDUL :: Mapa -> NBot -> Maybe Char

A função fugirDUL ajuda o bot a fugir das bombas, quando ele está indeciso entre down, up e left.

fugirDULAux :: Mapa -> [Coordenadas] -> Maybe Char

estudaDUL1 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDUL2 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDUL3 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDUL11 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDUL22 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDUL33 :: Mapa -> Coordenadas -> Maybe Char

fugirDUR :: Mapa -> NBot -> Maybe Char

A função fugirDUR ajuda o bot a fugir das bombas, quando ele está indeciso entre down, up e right.

fugirDURAux :: Mapa -> [Coordenadas] -> Maybe Char

estudaDUR1 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDUR2 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDUR3 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDUR11 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDUR22 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDUR33 :: Mapa -> Coordenadas -> Maybe Char

fugirULR :: Mapa -> NBot -> Maybe Char

A função fugirULR ajuda o bot a fugir das bombas, quando ele está indeciso entre up, left e right.

fugirULRAux :: Mapa -> [Coordenadas] -> Maybe Char

estudaULR1 :: Mapa -> [Coordenadas] -> Maybe Char

estudaULR2 :: Mapa -> [Coordenadas] -> Maybe Char

estudaULR3 :: Mapa -> [Coordenadas] -> Maybe Char

estudaULR11 :: Mapa -> [Coordenadas] -> Maybe Char

estudaULR22 :: Mapa -> [Coordenadas] -> Maybe Char

estudaULR33 :: Mapa -> Coordenadas -> Maybe Char

fugirRD :: Mapa -> NBot -> Maybe Char

A função fugirRD ajuda o bot a fugir das bombas, quando ele está indeciso entre right e down.

fugirRDAux :: Mapa -> [Coordenadas] -> Maybe Char

estudaRD1 :: Mapa -> [Coordenadas] -> Maybe Char

estudaRD2 :: Mapa -> [Coordenadas] -> Maybe Char

estudaRD11 :: Mapa -> [Coordenadas] -> Maybe Char

estudaRD22 :: Mapa -> Coordenadas -> Maybe Char

fugirLD :: Mapa -> NBot -> Maybe Char

A função fugirLD ajuda o bot a fugir das bombas, quando ele está indeciso entre left e down.

fugirLDAux :: Mapa -> [Coordenadas] -> Maybe Char

estudaLD1 :: Mapa -> [Coordenadas] -> Maybe Char

estudaLD2 :: Mapa -> [Coordenadas] -> Maybe Char

estudaLD11 :: Mapa -> [Coordenadas] -> Maybe Char

estudaLD22 :: Mapa -> Coordenadas -> Maybe Char

fugirDU :: Mapa -> NBot -> Maybe Char

A função fugirDU ajuda o bot a fugir das bombas, quando ele está indeciso entre down e up.

fugirDUAux :: Mapa -> [Coordenadas] -> Maybe Char

estudaDU1 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDU2 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDU11 :: Mapa -> [Coordenadas] -> Maybe Char

estudaDU22 :: Mapa -> Coordenadas -> Maybe Char

fugirLR :: Mapa -> NBot -> Maybe Char

A função fugirLR ajuda o bot a fugir das bombas, quando ele está indeciso entre left e right.

fugirLRAux :: Mapa -> [Coordenadas] -> Maybe Char

estudaLR1 :: Mapa -> [Coordenadas] -> Maybe Char

estudaLR2 :: Mapa -> [Coordenadas] -> Maybe Char

estudaLR11 :: Mapa -> [Coordenadas] -> Maybe Char

estudaLR22 :: Mapa -> Coordenadas -> Maybe Char

fugirUR :: Mapa -> NBot -> Maybe Char

A função fugirUR ajuda o bot a fugir das bombas, quando ele está indeciso entre up e right.

fugirURAux :: Mapa -> [Coordenadas] -> Maybe Char

estudaUR1 :: Mapa -> [Coordenadas] -> Maybe Char

estudaUR2 :: Mapa -> [Coordenadas] -> Maybe Char

estudaUR11 :: Mapa -> [Coordenadas] -> Maybe Char

estudaUR22 :: Mapa -> Coordenadas -> Maybe Char

fugirUL :: Mapa -> NBot -> Maybe Char

A função fugirUL ajuda o bot a fugir das bombas, quando ele está indeciso entre up e left.

fugirULAux :: Mapa -> [Coordenadas] -> Maybe Char

estudaUL1 :: Mapa -> [Coordenadas] -> Maybe Char

estudaUL2 :: Mapa -> [Coordenadas] -> Maybe Char

estudaUL11 :: Mapa -> [Coordenadas] -> Maybe Char

estudaUL22 :: Mapa -> [Coordenadas] -> Maybe Char

semPerigoLista :: Mapa -> [Coordenadas] -> Bool

A função semPerigoLista vê se uma lista das coordenadas de uma "direcao" para onde o bot pode ir não pertence as chamas das bombas plantadas.

paraOndePossoIrLista :: Mapa -> [Coordenadas] -> [Coordenadas]

semPerigo :: Mapa -> Coordenadas -> Bool

A função semPerigo vê se o bot está em perigo numa determinada coordenada.

paraOndePossoIr :: Mapa -> Coordenadas -> [Coordenadas]

A função paraOndePossoIr dá as 4 coordenadas (no máx.) que ele pode ir (tendo em atenção as pedras).

paraOndePossoIrAux :: Mapa -> [Coordenadas] -> [Coordenadas]

paraOndePossoIrTodos :: Mapa -> Coordenadas -> [Coordenadas]