Main

Documentation

main :: IO ()

Module : Main Description : Módulo Haskell que contém um exemplo de possível reação de um comando num estado de jogo Copyright : Catarina Machado; Joana Matos

Um módulo contendo funções recursivas e não recursivas com o intuito de devolver um mapa atualizado consoante comandos.

main : função de teste

move :: [String] -> Int -> Char -> [String]

move : função que devolve um novo estado de jogo consoante os comandos

type Coordenadas = (Int, Int)

coordenadasJ :: Char -> [String] -> String

coordenadasJ : função que devolve as coordenadas do jogador j

tiraSpace :: String -> String

tiraSpace : função que retira um espaço, se este for o último elemento de uma lista

retiraPUJogador :: String -> String

retiraPUJogador : função que retira power up do jogador

tirarNJogador :: String -> String

tirarNJogador : função auxiliar da função "coordenadasJ"

jogadorInicio :: Char -> [String] -> [Char]

jogadorInicio : função auxiliar da função "coordenadasJ"

coordJogador :: String -> (Int, Int)

coordJogador : função que devolve as coordenadas do jogador no tipo "coordenadas" definida anteriormente

coord1 :: String -> String

coord1 : função auxiliar da função "coordJogador"

coord2 :: String -> String

coord2 : função auxiliar da função "coordJogador"

whatIn :: [String] -> Char -> Char -> Char

whatIn : função que devolve a coordenada para onde um jogador j quer ir

whatInAux :: [String] -> (Int, Int) -> Char -> Char

whatInAux : função que auxilia na mostragem das coordenadas para onde um jogador j quer ir

andThen :: [String] -> Char -> Char -> Char -> [String]

andThen : função que vê o que acontece na coordenada consoante para onde o jogador vai

seraQTemPU :: [String] -> Char -> Char -> [String]

seraQTemPU : função que vê se o espaço tem power up ou não

interrogation :: (Int, Int) -> [(Int, Int)] -> [String] -> Char -> Char -> [String]

interrogation : função auxiliar de "seraQTemPu"

espacoVazio :: [String] -> Char -> Char -> [String]

espacoVazio : função que mostra o que acontece quando o jogador quer ir para um espaço vazio sem power ups

linhaNova :: [String] -> Char -> Char -> String

linhaNova : função que mostra as informações atuais do jogador

novaCoord :: (Int, Int) -> Char -> (Int, Int)

novaCoord = função que mostra as coordenadas em que o jogador fica

coordString :: (Int, Int) -> String

coordString : função que muda o tipo de coordenadas do jogador

howManyPU :: String -> String

howManyPU : função que dá informação de quais power ups o jogador tinha

powerupS :: String -> String

powerupS : função que coloca um espaço à frente se a lista não for vazia

espacoVazioCPU :: [String] -> Char -> Char -> [String]

espacoVazioCPU : função que acrescenta informação de power up à frente do jogador

eliminalinhaPU :: [String] -> Char -> Char -> [String]

eliminalinhaPU : função que elimina linha que dava a informação do power up

comparalinha :: [String] -> [String] -> Char -> Char -> [String]

comparalinha : função que apaga as linhas com informação de power ups

somaPU :: [String] -> Char -> Char -> [String]

somaPU : função que acrescenta "+" ou "!" à frente do jogador

eBombOrFlame :: [String] -> Char -> Char -> Bool

eBombOrFlame : função que testa se o jogador apanhou "+" ou "!"

acrescentaB :: String -> String

acrescentaB : função que acrescenta "+" à frente da nova coordenada do jogador

acrescentouB :: [String] -> Char -> Char -> String

acrescentouB : função que acrescenta ou não power ups

acrescentaF :: String -> String

acrescentaF : função que acrescenta "!" à frente da nova coordenada do jogador

acrescentouF :: [String] -> Char -> Char -> String

acrescentouF : função que acrescenta ou não power ups

substituiBomb :: [String] -> Char -> Char -> [String]

substituiBomb : função que atualiza informação

somaBomb :: [String] -> Char -> Char -> String

somaBomb :função que adiciona "+" na informação do jogador

somaBombAux :: String -> Bool

somaBombAux : função auxiliar de "somaBomb"

somaBombAux2 :: String -> String

somaBombAux2 : função auxiliar de "somaBombAux"

somaBombAux3 :: [String] -> Char -> Char -> String

somaBombAux3 : função auxiliar de "somaBombAux2"

somaBombAux4 :: String -> Bool

somaBombAux4 : função auxiliar de "somaBombAux3"

somaBombAux5 :: [String] -> Char -> Char -> String

somaBombAux5 : função auxiliar de "somaBombAux4"

substituiFlame :: [String] -> Char -> Char -> [String]

substituirFlame : função que atualiza a informação da flame

somaFlame :: [String] -> Char -> Char -> String

somaFlame : função que coloca "!" ou não à frente do jogador

novaInfoJ :: [String] -> Char -> Char -> String

novaInfoJ : função que devolve a informação do jogador j atualizada

coordenadasPU :: [String] -> [(Int, Int)]

coordenadasPU : função que devolve as coordenadas dos power ups no tipo [(Int,Int)]

coordenadassPU :: [String] -> [String]

coordenadassPU : função que dado um estado de jogo, devolve a informação relativa aos power ups

tirarPUeCV :: [String] -> [String]

tirarPUeCV : função auxiliar de "coordenadassPU"

auxTirar :: String -> String

auxTirar : função auxiliar de "tirarPUeCV"

powerUp :: [String] -> [String]

powerUp : função deixa a informação sobre power ups

poeBomba :: [String] -> Char -> [String]

poeBomba : função que vê em que síios podem ser postas as bombas

podeCBomba :: [String] -> Char -> Bool

podeCBomba : função que testa se a bomba pode ser colocada nessa posição

coordBombas :: [String] -> [String]

coordBombas : função que dá informação sobre as bombas colocadas

coordBombasAux :: [String] -> [String]

coordBombasAux : função auxiliar de "coordBombas"

linha1B :: String -> String

linha1B : função que dá as coordenadas de uma bomba já colocada

tiraRB :: String -> String

tiraRB : função auxiliar de "linha1B"

tiraRB1 :: String -> String

tiraRB1 : função auxiliar de "tiraRB1"

tira10 :: String -> String

tira10 : função que retira o raio da bomba à informação dada sobre a mesma

tiraBeE :: String -> String

tiraBeE : função que retira o primeiro e segundo elementos

procuraInfoBomba :: [String] -> [String]

procuraInfoBomba : função que dá informação das bombas que foram colocadas

podeColocarMaisB :: [String] -> Char -> Bool

podeColocarMaisB : função que apenas deixa o jogador plantar consoante os power ups bomb que tem

podeColocarMaisBaux :: Int -> Int -> Bool

podeColocarMaisBaux : função auxiliar de "podeColocarMaisB"

quantasPUB :: [String] -> Char -> Int

quantasPUB : funçãpo que mostra quantas bombas pode um jogador plantar

quantasPUBaux :: String -> Int

quantasPUBaux : função auxiliar de "quantasPUB"

quantasColocadas :: [String] -> Char -> Int

quantasColocadas : função que mostra as coordenadas dos jogadores e a quantidade de bombas que tem colocadas

qtsvezes :: Int -> [Int] -> Int

qtsvezes : função que dá o número de vezes que o jogador colocou

ordenaBB :: [String] -> Char -> [String]

ordenaBB : função que ordena a informação das bombas consoante as coordenadas

blibli :: [String] -> Char -> Int

bleble :: [String] -> Int

blabla :: (Int, Int) -> [(Int, Int)] -> Int

colocaLinha111 :: [String] -> Char -> [String]

colocaLinha111 : função que coloca a linha com informação da bomba no sitio certo

colocaLinha11 :: [String] -> Char -> String -> Int -> [String]

colocaLinha11 : função que auxilía na função "colocaLinha111"

coordBombasInt :: [String] -> [(Int, Int)]

coordBombasInt : função que dá as coordenadas das bombas que estão plantadas em [String] para [(Int,Int)]

coordBombasIntaux :: [String] -> [(Int, Int)]

coordBombasIntaux : função auxiliar na "coordBombasInt"

numerosIntJB :: [String] -> [Int]

numerosIntJB : função que passa a função "numerosJB" de [String] para [Int]

numerosIntJB1 :: [String] -> [Int]

numerosIntJB1 : função que dá auxílio na mudança de tipos da função "numerosJB"

numerosJB :: [String] -> [String]

numerosJB : função que coloca a lista com os números do jogadores que colocaram bombas

numerosJBAux :: [String] -> [String] -> [String]

numerosJBAux : função que auxilía na funcao "numerosJB"

poelinhaplanta :: [String] -> Char -> [String]

poelinhaplanta : função que coloca a "linhaplanta" no mapa (antes da informacao dos jogadores)

poelinhaplantaAux :: [String] -> Char -> String -> [String]

poelinhaplantaAux : função que auxilia a funcao "poelinhaplanta"

linhaplanta :: [String] -> Char -> String

linhaplanta : função que devolve uma linha nova com a informação sobre a colocação de uma bomba

raioBomb :: [String] -> Char -> String

raioBomb : função que altera o o tipo do raio de [String] para String

raioB :: [String] -> Char -> Int

raioB : função que fornece a quantidade de power ups do tipo flame

raioBAux :: String -> Int

raioBAux : função auxiliar que fornece o raio que terá a bomba, ou seja, a quantidade de power ups do tipo flame