Main

Documentation

encode :: [String] -> String

Main

Module : Main Description : Módulo Haskell que contém um exemplo de possível compressão e descompressão de um mapa de jogo Copyright : Catarina Machado; Joana Matos

Um módulo contendo funções recursivas e não recursivas com o intuito de devolver um mapa mais comprimido ou descomprimido, consoante o desejado.

encode : função que codifica o estado do jogo

decode :: String -> [String]

decode : função que descodifica o estado do jogo comrpimido anteriormente

main :: IO ()

main : função que comprime ou descomprime o mapa consoante o seu estado atual

separaMapa :: [String] -> [String]

separaMapa : função que mostra apenas o estado inicial referente ao mapa

separaCoordenadas :: [String] -> [String]

separaCoordenadas : função que mostra o estado das coordenadas do jogo

comprimir :: [String] -> [String]

comprimirMapa : função que decide o estado comprimido do jogo consoante o seu tamanho, distinguindo assim os mapas de dimensão 5 dos de dimensão maior que 5 pela razão de que um mapa de dimensão 5 tem possibilidade de maior compressão pois o seu formato será sempre igual independentemente do valor da semente

comprimirMapa :: [String] -> [String]

Exemplos de utilização:

>>> comprimir ["#####","#   #","# # #","#   #","#####","0 1 1"]
["v","01 1"]

>>> comprimir ["#########","#       #","# #?# # #","#     ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","! 5 5","* 7 7 1 1 10","0 3 3 ++","1 7 7"]
["7","1#?#1#1","5?1","?#1#1#?","1?2?2","1#?#?#1","2??3","!5 5","*7 7 1 1 10","03 3 ++","17 7"]

mapa5 :: [String] -> [String]

mapa5 : função que substitui um mapa de dimensão por um caractér

mapaMaior5 :: [String] -> [String]

mapaMaior5 : função que comprimirá um mapa de dimensão maior que 5

comprimirPedra :: [String] -> [String]

Uma das características que distingue a função mapa5 da função mapaMaior5: *calcula um mapa contraído de dimensão 5

pedrasLinha :: [String] -> [String]

pedrasLinha : função que retira as pedras (#) da primeira e última linhas, substituindo-as pelo vazio

primUltPedras :: [String] -> [String]

primUltPedras : função que retira o muro das outras linhas, ou seja, retira o primeiro e último elementos de cada linha

primPedra :: [String] -> [String]

primPedra : função que tira as pedras da primeira coluna

ultPedra :: [String] -> [String]

ultPedra : função que retira as pedras da última coluna

comprimirEspacos :: [String] -> [String]

comprimirEspacos : função que substitui os espaços pêlo seu número existencial

espacosNada :: [Char] -> Int -> [Char]

espacosNada : função que faz contagem dos espaços

comprimirCoordenadas :: [String] -> [String]

comprimirCoordenadas : função que comprime o espaço das coordenadas

tiraEspaCoord :: [String] -> [String]

tiraEspaCoord : função que retira o primeiro espaço existente nas coordenadas

tiraEspaCoordAux :: String -> String

tiraEspaCoordAux : função recursiva que retira o primeiro espaço de uma linha de coordenadas

descomprimir :: [Char] -> [Char]

descomprimir : função geral que descomprime o mapa juntamente com as coordenadas para o seu estado inicial

descomprimirMapa :: [Char] -> [Char]

descomprimirMapa : descomprimirá o mapa devolvendo o seu estado atual

descomprimir5 :: [Char] -> [Char]

descomprimir5 : função que descomprimirá apenas o estado do mapa

descomprimirEspacos :: [Char] -> [Char]

descomprimirEspacos : função que irá descomprimir os espaços inicialmente existentes no mapa

meterEspacos :: Char -> [Char]

meterEspacos : função que coloca espaços, substituindo-os pelos números previamente lá colocados

mapaString :: String -> [String]

mapaString : função que passa o mapa do tipo String para [String]

mapaChar :: [String] -> String

mapaChar : função que passa o mapa do tipo [String] para String

meterPedras :: String -> String

meterPedras : função geral do tipo String que coloca as pedras de novo no mapa

meterPedrasLString :: [String] -> [String]

meterPedrasLString : função do tipo [String] que coloca as pedras

meterPedrasC :: [String] -> [String]

meterPedrasC : função que coloca as pedras na primeira e última colunas

meterPedrasCAux :: String -> String

meterPedrasCAux : função auxiliar de "meterPedrasC"

meterPedrasL :: [String] -> [String]

meterPedrasL : função que coloca uma linha no fim e no início de pedras

descomprimirCoordenadas :: String -> String

descomprimirCoordenadas : função que nos fornece o estado inicial das coordenadas

descomprimirCoordenadasC :: String -> [String]

descomprimirCoordenadasC : função que nos fornece o estado inicial das coordenadas no tipo [String]

descomprimirCoordenadasAux :: [String] -> [String]

descomprimirCoordenadasAux : função auxiliar de "descomprimirCoordenadasC"