Main

Documentation

main :: IO ()

Module : Main Description : Módulo Haskell que contém um exemplo de possível criação de um mapa de jogo Copyright : Catarina Machado; Joana Matos

Um módulo contendo funções recursivas e não recursivas com o intuito de criar um mapa consoante uma dimensão e semente

main : função de teste para a função mapa

mapa :: Int -> Int -> [String]

mapa : função que dada uma dimensão e semente devolve uma lista do mapa

cbomb :: Int -> Int -> [String]

Exemplos de utilização:

>>> mapa 9 0 ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5"]

>>> mapa 5 0 ["#####","#   #","# # #","#   #","#####"]

cbomb : função que devolve as coordenadas de power ups do tipo bomb

cflame :: Int -> Int -> [String]

cflame : função que dá as coordenadas dos power ups tipo flame

coordenadaBomb :: [String] -> [String]

coordenadaBomb : função auxiliar da função "cbomb"

coordenadaFlame :: [String] -> [String]

coordenadaFlame : função auxiliar da função "cflame"

intToChar :: [(Int, Int)] -> [String]

intToChar : função que altera o tipo da função

coordenada1 :: (Int, Int) -> String

coordenada1 : função auxiliar da função "intToChar"

auxHB :: Int -> Int -> [Int]

auxHB : função auxiliar para "simplificar" coordenadaHB

auxHF :: Int -> Int -> [Int]

auxHF : função auxiliar para "simplificar" coordenadaHF

coordenadaHB :: [String] -> [Int]

coordenadaHF :: [String] -> [Int]

my_elemIndices2 :: Eq a => a -> [a] -> [Int]

coordenadaV :: Int -> [Int] -> [Int]

calculaV :: Int -> Int -> Int

calculaV : função auxiliar de "coordenadaV"

posicaoBomb :: Int -> Int -> [Int]

posicaoBomb : função que diz em que posição estão os power ups do tipo bomb

posicaoFlame :: Int -> Int -> [Int]

posicaoFlame : função que diz em que posição estão os power ups do tipo flame

my_elemIndices :: Eq a => a -> [a] -> [Int]

my_elemIndices : função auxiliar das funções "posicaoBomb" e "posicaoFlame"

substituir2 :: [Coordenadas] -> Int -> [Char] -> [Char] -> [Char] -> [Char]

substituir2 : função auxiliar de "posicaoFlame" e de "posicaoBomb"

porResultadoGama2 :: [Int] -> [Char]

porResultadoGama2 : função que altera "?" por "+" ou "!"

gerador :: Int -> Int -> [Int]

gerador : gera um determinado número de números aleatórios (consoante o número de células que inicialmente não podem ser célula vazia nem pedra)

agrupa :: Int -> [Char] -> [String]

agrupa : função que faz as linhas do mapa a partir da informação da função "substituir"

conta :: Int -> [Char] -> [Char]

conta : função auxiliar da função "agrupa"

type Coordenadas = (Int, Int)

substituir :: [Coordenadas] -> Int -> [Char] -> [Char] -> [Char] -> String

substituir : função que constrói o mapa tendo em conta onde deve haver pedras, espaços vazios ou pontos de interrogação

porResultadoGama :: [Int] -> [Char]

condicao :: [Coordenadas] -> Int -> [Char] -> [Char] -> [Char] -> [Char]

condicao : função auxiliar de "substituir" para um mapa de dimensão maior que 5

forPedra :: Coordenadas -> Int -> Bool

forPedra : função auxiliar de "condicao"

forVazia :: Coordenadas -> Int -> Bool

forVazia : função auxiliar de "condicao"

condicao5 :: [Coordenadas] -> Int -> [Char] -> [Char] -> [Char] -> [Char]

condicao5 : função auxiliar de "substituir" para um mapa de dimensão igual a 5

forPedra5 :: Coordenadas -> Int -> Bool

forPedra5 : função auxiliar de "condicao5"

forVazia5 :: Coordenadas -> Int -> Bool

forVazia5 : função auxiliar de "condicao5"

pedras :: Int -> [Coordenadas] -> [Char]

pedras : função que devolve as pedras existentes no mapa

dimensaoMore5 :: Int -> [Coordenadas] -> [Char]

dimensaoMore5 : função auxiliar de "pedras" para um mapa com dimensão maior que 5

dimensaoEqual5 :: Int -> [Coordenadas] -> [Char]

dimensaoEqual5 : função auxiliar de "pedras" para um mapa com dimensão 5

celulaVazia :: Int -> [Coordenadas] -> [Char]

celulaVazia : função que testa quantas células vazias tem o mapa

dimensaoMaiorQue5 :: Int -> [Coordenadas] -> [Char]

dimensaoMaiorQue5 : função auxiliar de "celulaVazia" para um mapa com dimensão maior que 5

dimensaoIgual5 :: Int -> [Coordenadas] -> [Char]

dimensaoIgual5 : função auxiliar de "celulaVazia" para um mapa de dimensão 5

contador :: Int -> [Coordenadas]

contador : função que devolve todas as coordenadas que existem no mapa

contadorX :: Int -> Int -> [Int]

contadorX : função auxiliar de "contador"

contadorY :: Int -> Int -> [Int]

contadorY : função auxiliar de "contador"

my_replicate :: Int -> [a] -> [a]

my_replicate : função auxiliar da função "contadorY"